Refactor ObjectDetectorFactory and implementations to appropriate layers
